fix memory problem

+++ utils/transformers_data_prep.py
@@ -78,7 +78,7 @@ def encode_examples(df, PATH, index, sample_weights=None, labels=None, forTest=F
 
 	with open(PATH + '/' + str(index) + '_input_mask.npy', 'wb') as filehandle:
 		# store the data as binary data stream
-		np.save(filehandle, np.asarray(_input['attention_mask'],  dtype=np.uint16))
+		np.save(filehandle, np.asarray(_input['attention_mask'],  dtype=np.uint8))
 
 	with open(PATH + '/' + str(index) + '_labels.npy', 'wb') as filehandle:
 		# store the data as binary data stream
